backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
basic-pitch==0.6.1
openai-whisper==20240930
demucs==4.0.1
music21==9.3.0
midiutil==1.2.1
numpy==1.26.4
scipy==1.14.1
soundfile==0.12.1
librosa==0.10.2.post1
